fix my code

+++ ft_puthex.c
@@ -6,13 +6,13 @@
 /*   By: yousenna <yousenna@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/22 11:51:05 by yousenna          #+#    #+#             */
-/*   Updated: 2025/11/24 19:01:06 by yousenna         ###   ########.fr       */
+/*   Updated: 2025/11/27 13:43:10 by yousenna         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft_printf.h"
+#include "ft_printf.h"
 
-int	ft_puthex(size_t nb, char c)
+int	ft_puthex(unsigned int nb, char c)
 {
 	char	*hex;
 	int		count;
